requirements.txt
Flask_Assets==0.12
psycopg2==2.7.3.2
Flask==0.12.2
fuzzywuzzy==0.15.1
cleanco==1.3
Flask_Cors==3.0.3
PyMySQL==0.7.11
Flask_Bcrypt==0.7.1
PyYAML==3.12
flask-mysql==1.4.0
gunicorn==19.7.1
newrelic==2.102.0.85
spacy
word2number
jellyfish
sklearn
pandas
scipy
google-cloud-vision
